+++ projects/_03_independent_cqc/_06_estimate_filled_posts/fargate/estimate_ind_cqc_filled_posts.py
@@ -1,4 +1,7 @@
+import polars as pl
+
 from polars_utils import utils
+from polars_utils.column_types import CategoricalColumnTypes as CatColType
 from projects._03_independent_cqc._06_estimate_filled_posts.fargate.utils.models.estimate_non_res_ct_filled_posts import (
     estimate_non_res_capacity_tracker_filled_posts,
 )
@@ -145,6 +148,24 @@ def main(
     lf = set_min_value(lf, IndCQC.estimate_filled_posts, 1.0)
 
     lf = estimate_non_res_capacity_tracker_filled_posts(lf)
+
+    lf = lf.with_columns(
+        pl.col(IndCQC.estimate_filled_posts_source).cast(
+            CatColType.EstimatesFilledPostSourceEnumType
+        ),
+        pl.col(
+            [
+                IndCQC.ascwds_filled_posts_dedup_clean,
+                IndCQC.ascwds_pir_merged,
+                IndCQC.care_home_model,
+                IndCQC.non_res_combined_model,
+                IndCQC.non_res_with_dormancy_model,
+                IndCQC.non_res_without_dormancy_model,
+                IndCQC.posts_rolling_average_model,
+                IndCQC.estimate_filled_posts,
+            ]
+        ).cast(pl.Float32),
+    )
 
     utils.sink_to_parquet(
         lf,